add zshrc to install command

+++ Makefile
@@ -6,7 +6,8 @@ test:
 	bash -c '. profile; bats tests'
 
 install: test
-	if [[ -e ~/.profile ]] || [[ -e ~/.profile.d ]]; then \
-		echo ~/.profile or ~/.profile.d exist; exit 1; fi
+	if [[ -e ~/.profile ]] || [[ -e ~/.profile.d ]] || [[ -e ~/.zshrc ]]; then \
+		echo ~/.profile, ~/.zshrc, or ~/.profile.d exist; exit 1; fi
 	ln -s "$$(pwd)/profile" ~/.profile
+	ln -s "$$(pwd)/zshrc" ~/.zshrc
 	ln -s "$$(pwd)/profile.d" ~/.profile.d
